+++ src/main/java/ai/preferred/venom/FatalHandlerException.java
@@ -1,9 +1,25 @@
+/*
+ * Copyright 2018 Preferred.AI
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package ai.preferred.venom;
 
 /**
  * This class defines fatal runtime exception for {@link Handler}.
  * If {@link Handler#handle(ai.preferred.venom.request.Request,
- * ai.preferred.venom.response.VResponse, ai.preferred.venom.Scheduler,
+ * ai.preferred.venom.response.VResponse, ai.preferred.venom.job.Scheduler,
  * Session, Worker)} encounters unexpected situation, it can signal
  * {@link Crawler} to stop execution by throwing this exception.
  *
