+++ src/main/java/ai/preferred/venom/Handler.java
@@ -16,6 +16,7 @@
 
 package ai.preferred.venom;
 
+import ai.preferred.venom.job.Scheduler;
 import ai.preferred.venom.request.Request;
 import ai.preferred.venom.response.VResponse;
