+++ src/main/java/ai/preferred/venom/job/QueueScheduler.java
@@ -17,8 +17,6 @@
 package ai.preferred.venom.job;
 
 
-import ai.preferred.venom.Scheduler;
-
 import java.util.concurrent.BlockingQueue;
 
 /**
